// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
      employees: {
        Row: {
          bank_account_no: string | null
          bank_name: string | null
          basic_salary: number | null
          branch_name: string | null
          date_of_birth: string | null
          company: string | null
          code: string | null
          created_at: string
          department: string | null
          designation: string | null
          esi_deduction: number | null
          esic_number: string | null
          hra: number | null
          id: string
          ifsc_code: string | null
          name: string
          other_allowances: number | null
          other_deductions: number | null
          pf_deduction: number | null
          pf_number: string | null
          professional_tax: number | null
          uan_number: string | null
          updated_at: string
          work_location: string | null
        }
        Insert: {
          bank_account_no?: string | null
          bank_name?: string | null
          basic_salary?: number | null
          branch_name?: string | null
          date_of_birth?: string | null
          company?: string | null
          code?: string | null
          created_at?: string
          department?: string | null
          designation?: string | null
          esi_deduction?: number | null
          esic_number?: string | null
          hra?: number | null
          id?: string
          ifsc_code?: string | null
          name: string
          other_allowances?: number | null
          other_deductions?: number | null
          pf_deduction?: number | null
          pf_number?: string | null
          professional_tax?: number | null
          uan_number?: string | null
          updated_at?: string
          work_location?: string | null
        }
        Update: {
          bank_account_no?: string | null
          bank_name?: string | null
          basic_salary?: number | null
          branch_name?: string | null
          date_of_birth?: string | null
          company?: string | null
          code?: string | null
          created_at?: string
          department?: string | null
          designation?: string | null
          esi_deduction?: number | null
          esic_number?: string | null
          hra?: number | null
          id?: string
          ifsc_code?: string | null
          name?: string
          other_allowances?: number | null
          other_deductions?: number | null
          pf_deduction?: number | null
          pf_number?: string | null
          professional_tax?: number | null
          uan_number?: string | null
          updated_at?: string
          work_location?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never